+++ common/src/main/java/net/smackplays/smacksutil/platform/services/IVeinMiner.java
@@ -6,15 +6,22 @@ import net.minecraft.client.Minecraft;
 import net.minecraft.client.gui.screens.Screen;
 import net.minecraft.core.BlockPos;
 import net.minecraft.network.chat.Component;
+import net.minecraft.server.level.ServerLevel;
 import net.minecraft.server.level.ServerPlayer;
 import net.minecraft.util.Mth;
+import net.minecraft.world.entity.item.ItemEntity;
 import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.item.Item;
 import net.minecraft.world.item.ItemStack;
 import net.minecraft.world.item.TieredItem;
 import net.minecraft.world.level.Level;
 import net.minecraft.world.level.block.Block;
+import net.minecraft.world.level.block.Blocks;
+import net.minecraft.world.level.block.entity.BlockEntity;
 import net.minecraft.world.level.block.state.BlockState;
+import net.minecraft.world.level.storage.loot.LootParams;
+import net.minecraft.world.level.storage.loot.parameters.LootContextParams;
+import net.minecraft.world.phys.Vec3;
 import net.minecraft.world.phys.shapes.Shapes;
 import net.minecraft.world.phys.shapes.VoxelShape;
 import net.smackplays.smacksutil.platform.Services;
@@ -116,7 +123,7 @@ public abstract class IVeinMiner {
         boolean mainHandIsTool = TieredItem.class.isAssignableFrom(mainHand.getClass());
         if (player.isCreative()) isCreative = true;
 
-        toBreak = getBlocks(world, player, sourcePos);
+        toBreak = (ArrayList<BlockPos>) getBlocks(world, player, sourcePos).clone();
 
         int maxDMG = mainHandStack.getMaxDamage();
 
@@ -132,8 +139,11 @@ public abstract class IVeinMiner {
 
             boolean canHarvest = (player.hasCorrectToolForDrops(currBlockState) || player.isCreative());
             if (canHarvest) {
-                world.destroyBlock(curr, !isCreative);
+                //world.destroyBlock(curr, !isCreative);
+                world.setBlockAndUpdate(curr, Blocks.AIR.defaultBlockState());
                 if (!isCreative) {
+                    BlockEntity currBlockEntity = currBlockState.hasBlockEntity() ? world.getBlockEntity(curr) : null;
+                    Block.dropResources(currBlockState, world, curr, currBlockEntity, null, ItemStack.EMPTY);
                     if (mainHandStack.isDamageableItem()) {
                         mainHandStack.hurt(1, player.getRandom(), (ServerPlayer) player);
                     }