+++ common/src/main/java/net/smackplays/smacksutil/util/PlayerComparator.java
@@ -1,4 +1,4 @@
-package net.smackplays.smacksutil.veinminer.modes;
+package net.smackplays.smacksutil.util;
 
 
 import net.minecraft.core.BlockPos;
@@ -6,19 +6,19 @@ import net.minecraft.world.entity.player.Player;
 
 import java.util.Comparator;
 
-public class BlockPosComparator implements Comparator<BlockPos> {
+public class PlayerComparator implements Comparator<BlockPos> {
 
     Player player;
 
-    public BlockPosComparator(Player playerIn) {
+    public PlayerComparator(Player playerIn) {
         player = playerIn;
     }
 
     @Override
     public int compare(BlockPos o1, BlockPos o2) {
         if (player == null) return 0;
-        double o1ToPlayer = o1.distToCenterSqr(player.position());
-        double o2ToPlayer = o2.distToCenterSqr(player.position());
+        double o1ToPlayer = o1.distToCenterSqr(player.blockPosition().getCenter());
+        double o2ToPlayer = o2.distToCenterSqr(player.blockPosition().getCenter());
         if (o1ToPlayer < o2ToPlayer) return -1;
         else if (o1ToPlayer == o2ToPlayer) return 0;
         else if (o1ToPlayer > o2ToPlayer) return 1;