+++ common/src/main/java/net/smackplays/smacksutil/veinminer/modes/Mineshaft.java
@@ -6,6 +6,7 @@ import net.minecraft.tags.TagKey;
 import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.level.Level;
 import net.minecraft.world.level.block.Block;
+import net.smackplays.smacksutil.util.BlockPosComparator;
 import net.smackplays.smacksutil.util.ModTags;
 
 import java.util.ArrayList;
@@ -22,6 +23,7 @@ public class Mineshaft extends VeinMode {
         if (world == null || player == null || sourcePos == null) return (ArrayList<BlockPos>) toBreak.clone();
         oldToBreak = (ArrayList<BlockPos>) toBreak.clone();
         toBreak.clear();
+        toCheck.clear();
         checked.clear();
         Block toMatch = world.getBlockState(sourcePos).getBlock();
 
@@ -41,7 +43,7 @@ public class Mineshaft extends VeinMode {
 
         mineshaft(sourcePos, player.getDirection(), radius, player, world, isExactMatch, toMatch, tag);
 
-        toBreak.sort(new BlockPosComparator(player));
+        toBreak.sort(new BlockPosComparator(sourcePos));
 
         oldToBreak = (ArrayList<BlockPos>) toBreak.clone();
         oldRadius = radius;