+++ common/src/main/java/net/smackplays/smacksutil/veinminer/modes/Shapeless.java
@@ -6,6 +6,7 @@ import net.minecraft.world.entity.player.Player;
 import net.minecraft.world.level.Level;
 import net.minecraft.world.level.block.Block;
 import net.smackplays.smacksutil.platform.Services;
+import net.smackplays.smacksutil.util.BlockPosComparator;
 import net.smackplays.smacksutil.util.ModTags;
 
 import java.util.ArrayList;
@@ -21,6 +22,7 @@ public class Shapeless extends VeinMode {
     public ArrayList<BlockPos> getBlocks(Level world, Player player, BlockPos sourcePos, int radius, boolean isExactMatch) {
         if (world == null || player == null || sourcePos == null) return (ArrayList<BlockPos>) toBreak.clone();
         toBreak.clear();
+        toCheck.clear();
         checked.clear();
         Block toMatch = world.getBlockState(sourcePos).getBlock();
 
@@ -37,7 +39,7 @@ public class Shapeless extends VeinMode {
         }
         shapeless(new BlockPos(sourcePos), world, player, radius, isExactMatch, toMatch, tag);
 
-        toBreak.sort(new BlockPosComparator(player));
+        toBreak.sort(new BlockPosComparator(sourcePos));
 
         oldToBreak = (ArrayList<BlockPos>) toBreak.clone();
         oldRadius = radius;
@@ -48,9 +50,9 @@ public class Shapeless extends VeinMode {
         return (ArrayList<BlockPos>) toBreak.clone();
     }
 
-    private void shapeless(BlockPos curr, Level world, Player player, int radius, boolean isExactMatch, Block toMatch, TagKey<Block> tag) {
-        toCheck.add(curr);
-        while (!toCheck.isEmpty() && toCheck.get(0).distToCenterSqr(curr.getCenter()) < radius) {
+    private void shapeless(BlockPos sourcePos, Level world, Player player, int radius, boolean isExactMatch, Block toMatch, TagKey<Block> tag) {
+        toCheck.add(sourcePos);
+        while (!toCheck.isEmpty()) {
             BlockPos currPos = toCheck.get(0);
             if (checkMatch(isExactMatch, currPos, world, player, toMatch, tag)) {
                 toBreak.add(currPos);
@@ -60,11 +62,16 @@ public class Shapeless extends VeinMode {
             toCheck.remove(currPos);
             // remove duplicates
             ArrayList<BlockPos> newList = new ArrayList<>();
-            for (BlockPos p : toCheck) {
-                if (!newList.contains(p) && !toBreak.contains(p)) newList.add(p);
+            for (BlockPos p : (ArrayList<BlockPos>)toCheck.clone()) {
+                if (!newList.contains(p) && !toBreak.contains(p)
+                        && !((p.getX() > sourcePos.getX() + radius) || (p.getX() < sourcePos.getX() - radius))
+                        && !((p.getY() > sourcePos.getY() + radius) || (p.getY() < sourcePos.getY() - radius))
+                        && !((p.getZ() > sourcePos.getZ() + radius) || (p.getZ() < sourcePos.getZ() - radius))){
+                    newList.add(p);
+                }
             }
             toCheck = newList;
-            toCheck.sort(new BlockPosComparator(player));
+            toCheck.sort(new BlockPosComparator(sourcePos));
         }
     }